Amélioration de la documentation sur le comportement de l'interface verrouillée. Changement des méthodes de gestion des éléments interactifs pour utiliser un `unordered_map`, augmentant ainsi l'efficacité. Ajustements dans la gestion des événements pour assurer une détection correcte des éléments interactifs.
Changements mineurs de style et de formatage pour améliorer la lisibilité. Modification de la fonction `addInteractive` pour accepter un `std::string` au lieu d'un `std::string_view`. Mise à jour de la logique de détection des éléments survolés pour garantir une identification correcte et une gestion appropriée des interactions.

+++ src/GUI/AdvancedInterface.cpp
@@ -89,7 +89,7 @@ void hideSlider(InteractiveInterface* gui, const std::string& identifier, bool h
 	textPtr->hide   = hide;
 }
 
-float moveSlider(InteractiveInterface* gui, const std::string& identifier, float yPos, int intervals, const GrowthSliderFunction& growth, const UserFunction& user)
+double moveSlider(InteractiveInterface* gui, const std::string& identifier, double yPos, int intervals, const GrowthSliderFunction& growth, const UserFunction& user)
 {
 	ENSURE_VALID_PTR(gui, "The gui was nullptr when the function moveSlider was called");
 
@@ -100,8 +100,8 @@ float moveSlider(InteractiveInterface* gui, const std::string& identifier, float
 	if (backgroundSlider == nullptr || cursorSlider == nullptr)
 		throw std::invalid_argument{ "The slider with the identifier " + identifier + " does not exist." };
 
-	const float bias{ backgroundSlider->getSprite().getGlobalBounds().position.y };
-	const float length{ backgroundSlider->getSprite().getGlobalBounds().size.y };
+	const double bias{ backgroundSlider->getSprite().getGlobalBounds().position.y };
+	const double length{ backgroundSlider->getSprite().getGlobalBounds().size.y };
 	yPos = std::clamp(yPos, bias, bias + length) - bias;
 
 	if (intervals >= 0)
@@ -110,12 +110,12 @@ float moveSlider(InteractiveInterface* gui, const std::string& identifier, float
 		yPos = (length * round((yPos * intervals) / length)) / intervals; // Rounding to the nearest interval position.
 	}
 
-	const float value{ growth(1 - ((yPos) / length)) }; // Get the value of the slider.
+	const double value{ growth(1 - ((yPos) / length)) }; // Get the value of the slider.
 	
 	yPos += bias;
-	cursorSlider->setPosition(sf::Vector2f{ cursorSlider->getSprite().getPosition().x, yPos });
+	cursorSlider->setPosition(sf::Vector2f{ cursorSlider->getSprite().getPosition().x, static_cast<float>(yPos) });
 	textSlider->setContent(value);
-	textSlider->setPosition(sf::Vector2f{ textSlider->getText().getPosition().x, yPos }); // Aligning the text with the cursor.
+	textSlider->setPosition(sf::Vector2f{ textSlider->getText().getPosition().x, static_cast<float>(yPos) }); // Aligning the text with the cursor.
 
 	if (user != nullptr)
 		user(value); // Call the function associated with the slider.
